+++ chat-widget/src/components/livechatwidget/interfaces/ILiveChatWidgetProps.ts
@@ -3,7 +3,6 @@ import { IAppInsightsConfig } from "../../../common/telemetry/interfaces/IAppIns
 import { IAudioNotificationProps } from "../../footerstateful/audionotificationstateful/interfaces/IAudioNotificationProps";
 import { ICallingContainerProps } from "@microsoft/omnichannel-chat-components/lib/types/components/callingcontainer/interfaces/ICallingContainerProps";
 import { IChatButtonProps } from "@microsoft/omnichannel-chat-components/lib/types/components/chatbutton/interfaces/IChatButtonProps";
-import { ICitationPaneProps } from "@microsoft/omnichannel-chat-components/lib/types/components/citationpane/interfaces/ICitationPaneProps";
 import { ICitationPaneStatefulProps } from "../../citationpanestateful/interfaces/ICitationPaneStatefulProps";
 import { IConfirmationPaneStatefulProps } from "../../confirmationpanestateful/interfaces/IConfirmationPaneStatefulProps";
 import { IContextDataStore } from "../../../common/interfaces/IContextDataStore";
@@ -79,5 +78,4 @@ export interface ILiveChatWidgetProps {
     featureConfigProps?: IFeatureConfigProps;
     appInsightsConfig?: IAppInsightsConfig;
     persistentChatHistoryProps?: IPersistentChatHistoryProps;
-    citationProp?: ICitationPaneProps;
 }